+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'book_list.dart';
-import 'book_add_manually.dart';
+import 'book_add_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -17,28 +17,8 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
         appBar: AppBar(title: Text('あなたの本棚')),
         body: BodyWidget(),
-        floatingActionButton: FloatingButton(),
       ),
     );
   }
 }
 
-class FloatingButton extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomRight,
-      child: new FloatingActionButton(
-        child: new Icon(Icons.add_box),
-        onPressed: () => {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-            builder: (context) => BookAddScreen(),
-            ),
-          )
-        }
-      ),
-    );
-  }
-}
